+++ syft/frameworks/torch/nn/__init__.py
@@ -6,6 +6,6 @@ from syft.frameworks.torch.nn.rnn import GRUCell
 from syft.frameworks.torch.nn.rnn import LSTMCell
 
 from syft.frameworks.torch.nn.conv import Conv2d
+from syft.frameworks.torch.nn import functional
 
 from syft.frameworks.torch.nn.pool import AvgPool2d
-from syft.frameworks.torch.nn.functional import dropout
